requirements.txt
Pillow
smtplib
tk
